chat-youtube-video/requirements.txt
pytube
openai-whisper
streamlit
openai
